+++ src/main/java/ladder/strategy/StepRandomStrategy.java
@@ -1,4 +1,4 @@
-package ladder.service;
+package ladder.strategy;
 
 import java.util.Random;
