+++ src/test/java/ladder/domain/ladder/FootStepTest.java
@@ -1,6 +1,6 @@
 package ladder.domain.ladder;
 
-import ladder.service.StepStrategy;
+import ladder.strategy.StepStrategy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
